cleanup of config.c: move more code to other modules like unit.c and faction.c
