+++ src/kernel/faction.c
@@ -45,13 +45,15 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <util/lists.h>
 #include <util/language.h>
 #include <util/log.h>
-#include <quicklist.h>
+#include <util/parser.h>
 #include <util/resolve.h>
 #include <util/rng.h>
 #include <util/variant.h>
 #include <util/unicode.h>
+
 #include <attributes/otherfaction.h>
 
+#include <quicklist.h>
 #include <storage.h>
 
 /* libc includes */
@@ -150,6 +152,11 @@ faction *findfaction(int n)
 {
     faction *f = ffindhash(n);
     return f;
+}
+
+faction *getfaction(void)
+{
+    return findfaction(getid());
 }
 
 void set_show_item(faction * f, const struct item_type *itype)
@@ -803,3 +810,36 @@ int max_magicians(const faction * f)
     return m;
 }
 
+#define DMAXHASH 7919
+typedef struct dead {
+    struct dead *nexthash;
+    faction *f;
+    int no;
+} dead;
+
+static dead *deadhash[DMAXHASH];
+
+void dhash(int no, faction * f)
+{
+    dead *hash = (dead *)calloc(1, sizeof(dead));
+    dead *old = deadhash[no % DMAXHASH];
+    hash->no = no;
+    hash->f = f;
+    deadhash[no % DMAXHASH] = hash;
+    hash->nexthash = old;
+}
+
+faction *dfindhash(int no)
+{
+    dead *old;
+
+    if (no < 0)
+        return 0;
+
+    for (old = deadhash[no % DMAXHASH]; old; old = old->nexthash) {
+        if (old->no == no) {
+            return old->f;
+        }
+    }
+    return 0;
+}
